+++ src/workflowOptions.ts
@@ -47,7 +47,7 @@ export const workflowOptions = {
     type: "select",
     name: "network",
     message: "Choose your network",
-    choices(prev, values, prompt) {
+    choices(prev) {
       if (prev.path === "boilerplate-template") {
         return [
           { title: "Mainnet", value: "mainnet" },
@@ -62,5 +62,11 @@ export const workflowOptions = {
     },
     initial: 0,
     hint: "- You can change this later",
+  },
+  analytics: {
+    type: "confirm",
+    name: "telemetry",
+    message: "Help us improve create-aptos-dapp by collection anonymous data",
+    initial: true,
   },
 };